Transition to condensed/uncondensed vs. interior/trace
This coincides with an added capability for the application code
to specify variables for which there should be no condensation.
I am now seeing a successful solve for `grid_size = 1`, but as soon
as I went to `grid_size = 2`, then the local condensed blocks
became singular again.

A source of comfort (maybe) is that looking back at the devel/master
implementation of the HDG NS example, the matrices there appear to
be identical (within permutations of rows columns for different local
numbering) and so the determinants there are identical: ~6e-14. I
guess maybe just the `inverse()` method is more stable than the LU
decomposition? Either way it seems like we need to go to an implemetation
like in Nguyen where there are both interior and trace pressure dofs
or just update to an IP-HDG method like Sander's.