Don't trust determinant to tell you if matrix is singular
I feel really dumb. I didn't know this relationship:

> Absolute value of the determinant of square matrix A is equal to
> the product of its singular values

So if all the singular values are small then the deterimant is very
small. However, the condition number is the ratio of largest and
smallest singular values, so you could have a condition number of 1
and a determinant that is miniscule